+++ node/Cargo.toml
@@ -37,7 +37,6 @@ async-trait = "0.1.80"
 beetswap = "0.1.1"
 cid = { version = "0.11.1", features = ["serde-codec"] }
 dashmap = "5.5.3"
-either = "1.13.0"
 futures = "0.3.30"
 instant = "0.1.13"
 prost = "0.12.6"
