+++ node/src/syncer.rs
@@ -687,6 +687,77 @@ mod tests {
     use crate::test_utils::{async_test, gen_filled_store, MockP2pHandle};
     use celestia_types::test_utils::ExtendedHeaderGenerator;
 
+    #[test]
+    fn calculate_range_to_fetch_test_header_limit() {
+        let head_height = 1024;
+        let ranges = [256..=512];
+
+        let fetch_range = calculate_range_to_fetch(head_height, &ranges, 16);
+        assert_eq!(fetch_range, 513..=528);
+
+        let fetch_range = calculate_range_to_fetch(head_height, &ranges, 511);
+        assert_eq!(fetch_range, 513..=1023);
+        let fetch_range = calculate_range_to_fetch(head_height, &ranges, 512);
+        assert_eq!(fetch_range, 513..=1024);
+        let fetch_range = calculate_range_to_fetch(head_height, &ranges, 513);
+        assert_eq!(fetch_range, 513..=1024);
+
+        let fetch_range = calculate_range_to_fetch(head_height, &ranges, 1024);
+        assert_eq!(fetch_range, 513..=1024);
+    }
+
+    #[test]
+    fn calculate_range_to_fetch_empty_store() {
+        let fetch_range = calculate_range_to_fetch(1, &[], 100);
+        assert_eq!(fetch_range, 1..=1);
+
+        let fetch_range = calculate_range_to_fetch(100, &[], 10);
+        assert_eq!(fetch_range, 1..=10);
+
+        let fetch_range = calculate_range_to_fetch(100, &[], 50);
+        assert_eq!(fetch_range, 1..=50);
+    }
+
+    #[test]
+    fn calculate_range_to_fetch_fully_synced() {
+        let fetch_range = calculate_range_to_fetch(1, &[1..=1], 100);
+        assert!(fetch_range.is_empty());
+
+        let fetch_range = calculate_range_to_fetch(100, &[1..=100], 10);
+        assert!(fetch_range.is_empty());
+
+        let fetch_range = calculate_range_to_fetch(100, &[1..=100], 10);
+        assert!(fetch_range.is_empty());
+    }
+
+    #[test]
+    fn calculate_range_to_fetch_caught_up() {
+        let head_height = 4000;
+
+        let fetch_range = calculate_range_to_fetch(head_height, &[3000..=4000], 500);
+        assert_eq!(fetch_range, 2500..=2999);
+        let fetch_range = calculate_range_to_fetch(head_height, &[500..=1000, 3000..=4000], 500);
+        assert_eq!(fetch_range, 2500..=2999);
+        let fetch_range = calculate_range_to_fetch(head_height, &[2500..=2800, 3000..=4000], 500);
+        assert_eq!(fetch_range, 2801..=2999);
+        let fetch_range = calculate_range_to_fetch(head_height, &[2500..=2800, 3000..=4000], 500);
+        assert_eq!(fetch_range, 2801..=2999);
+        let fetch_range = calculate_range_to_fetch(head_height, &[300..=4000], 500);
+        assert_eq!(fetch_range, 1..=299);
+    }
+
+    #[test]
+    fn calculate_range_to_fetch_catching_up() {
+        let head_height = 4000;
+
+        let fetch_range = calculate_range_to_fetch(head_height, &[2000..=3000], 500);
+        assert_eq!(fetch_range, 3001..=3500);
+        let fetch_range = calculate_range_to_fetch(head_height, &[2000..=3500], 500);
+        assert_eq!(fetch_range, 3501..=4000);
+        let fetch_range = calculate_range_to_fetch(head_height, &[1..=2998, 3000..=3800], 500);
+        assert_eq!(fetch_range, 3801..=4000);
+    }
+
     #[async_test]
     async fn init_without_genesis_hash() {
         let events = EventChannel::new();
